+++ packages/core/src/blocks/FileBlockContent/helpers/parse/parseEmbedElement.ts
@@ -1,0 +1,5 @@
+export const parseEmbedElement = (embedElement: HTMLEmbedElement) => {
+  const url = embedElement.src || undefined;
+
+  return { url };
+};
