+++ packages/core/src/blocks/ImageBlockContent/parseImageElement.ts
