+++ packages/react/src/blocks/FileBlockContent/helpers/render/ResizableFileBlockWrapper.tsx
@@ -1,129 +1,9 @@
 import { FileBlockConfig } from "@blocknote/core";
-import {
-  CSSProperties,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
-import { RiFile2Line } from "react-icons/ri";
+import { ReactNode, useCallback, useEffect, useRef, useState } from "react";
 
-import { useUploadLoading } from "../../hooks/useUploadLoading.js";
-import { useDictionary } from "../../i18n/dictionary.js";
-import { ReactCustomBlockRenderProps } from "../../schema/ReactBlockSpec.js";
-
-export const AddFileButton = (
-  props: Omit<
-    ReactCustomBlockRenderProps<FileBlockConfig, any, any>,
-    "contentRef"
-  > & {
-    buttonText?: string;
-    buttonIcon?: ReactNode;
-  }
-) => {
-  const dict = useDictionary();
-
-  // Prevents focus from moving to the button.
-  const addFileButtonMouseDownHandler = useCallback(
-    (event: React.MouseEvent) => {
-      event.preventDefault();
-    },
-    []
-  );
-  // Opens the file toolbar.
-  const addFileButtonClickHandler = useCallback(() => {
-    props.editor.dispatch(
-      props.editor._tiptapEditor.state.tr.setMeta(
-        props.editor.filePanel!.plugin,
-        {
-          block: props.block,
-        }
-      )
-    );
-  }, [props.block, props.editor]);
-
-  return (
-    <div
-      className={"bn-add-file-button"}
-      onMouseDown={addFileButtonMouseDownHandler}
-      onClick={addFileButtonClickHandler}>
-      <div className={"bn-add-file-button-icon"}>
-        {props.buttonIcon || <RiFile2Line size={24} />}
-      </div>
-      <div className={"bn-add-file-button-text"}>
-        {props.buttonText || dict.file_blocks.file.add_button_text}
-      </div>
-    </div>
-  );
-};
-
-export const FileNameWithIcon = (
-  props: Omit<
-    ReactCustomBlockRenderProps<FileBlockConfig, any, any>,
-    "editor" | "contentRef"
-  >
-) => (
-  <div
-    className={"bn-file-name-with-icon"}
-    contentEditable={false}
-    draggable={false}>
-    <div className={"bn-file-icon"}>
-      <RiFile2Line size={24} />
-    </div>
-    <p className={"bn-file-name"}>{props.block.props.name}</p>
-  </div>
-);
-
-export const FileBlockWrapper = (
-  props: Omit<
-    ReactCustomBlockRenderProps<FileBlockConfig, any, any>,
-    "contentRef"
-  > & {
-    buttonText?: string;
-    buttonIcon?: ReactNode;
-    children?: ReactNode;
-  } & {
-    // These props & the `forwardRef` are just here so we can reuse this
-    // component in `ResizableFileBlockWrapper`.
-    onMouseEnter?: () => void;
-    onMouseLeave?: () => void;
-    style?: CSSProperties;
-  }
-) => {
-  const showLoader = useUploadLoading(props.block.id);
-
-  return (
-    <div
-      className={"bn-file-block-content-wrapper"}
-      onMouseEnter={props.onMouseEnter}
-      onMouseLeave={props.onMouseLeave}
-      style={props.style}>
-      {showLoader ? (
-        // Show loader while a file is being uploaded.
-        <div className={"bn-file-loading-preview"}>Loading...</div>
-      ) : props.block.props.url === "" ? (
-        // Show the add file button if the file has not been uploaded yet.
-        <AddFileButton {...props} />
-      ) : (
-        // Show the file preview, or the file name and icon.
-        <>
-          {props.block.props.showPreview === false || !props.children ? (
-            // Show file name and icon.
-            <FileNameWithIcon {...props} />
-          ) : (
-            // Show preview.
-            props.children
-          )}
-          {props.block.props.caption && (
-            // Show the caption if there is one.
-            <p className={"bn-file-caption"}>{props.block.props.caption}</p>
-          )}
-        </>
-      )}
-    </div>
-  );
-};
+import { useUploadLoading } from "../../../../hooks/useUploadLoading.js";
+import { ReactCustomBlockRenderProps } from "../../../../schema/ReactBlockSpec.js";
+import { FileBlockWrapper } from "./FileBlockWrapper.js";
 
 export const ResizableFileBlockWrapper = (
   props: Omit<
@@ -287,23 +167,3 @@ export const ResizableFileBlockWrapper = (
     </FileBlockWrapper>
   );
 };
-
-export const LinkWithCaption = (props: {
-  caption: string;
-  children: ReactNode;
-}) => (
-  <div>
-    {props.children}
-    <p>{props.caption}</p>
-  </div>
-);
-
-export const FigureWithCaption = (props: {
-  caption: string;
-  children: ReactNode;
-}) => (
-  <figure>
-    {props.children}
-    <figcaption>{props.caption}</figcaption>
-  </figure>
-);
