+++ packages/react/src/blocks/FileBlockContent/helpers/toExternalHTML/FigureWithCaption.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+export const FigureWithCaption = (props: {
+  caption: string;
+  children: ReactNode;
+}) => (
+  <figure>
+    {props.children}
+    <figcaption>{props.caption}</figcaption>
+  </figure>
+);
