+++ packages/react/src/blocks/FileBlockContent/helpers/toExternalHTML/LinkWithCaption.tsx
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+export const LinkWithCaption = (props: {
+  caption: string;
+  children: ReactNode;
+}) => (
+  <div>
+    {props.children}
+    <p>{props.caption}</p>
+  </div>
+);
