+++ packages/react/src/index.ts
@@ -7,7 +7,12 @@ export * from "./i18n/dictionary.js";
 
 export * from "./blocks/AudioBlockContent/AudioBlockContent.js";
 export * from "./blocks/FileBlockContent/FileBlockContent.js";
-export * from "./blocks/FileBlockContent/fileBlockHelpers.js";
+export * from "./blocks/FileBlockContent/helpers/render/AddFileButton.js";
+export * from "./blocks/FileBlockContent/helpers/render/FileBlockWrapper.js";
+export * from "./blocks/FileBlockContent/helpers/render/FileNameWithIcon.js";
+export * from "./blocks/FileBlockContent/helpers/render/ResizableFileBlockWrapper.js";
+export * from "./blocks/FileBlockContent/helpers/toExternalHTML/FigureWithCaption.js";
+export * from "./blocks/FileBlockContent/helpers/toExternalHTML/LinkWithCaption.js";
 export * from "./blocks/FileBlockContent/useResolveUrl.js";
 export * from "./blocks/ImageBlockContent/ImageBlockContent.js";
 export * from "./blocks/VideoBlockContent/VideoBlockContent.js";
